pyproject.toml
[project]
name = "ocr-app"
version = "0.1.0"
description = ""
authors = [
    {name = "Zh3ka21",email = "puzanov_zhenechk@ukr.net"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "alembic (>=1.16.2,<2.0.0)",
    "paddleocr (>=3.1.0,<4.0.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "fastapi (>=0.115.14,<0.116.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "cryptography (>=45.0.4,<46.0.0)"
]

[tool.poetry]
packages = [{include = "ocr_app", from = "src"}]


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.16.1"
ruff = "^0.12.1"
pre-commit = "^4.2.0"
python-multipart = "^0.0.20"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
